Add files via upload
The project is about bank loan and how high or low interest rate influence bank customers to get a loan or not. A 12% double digit interest rate swayed 39,000 bank customers to applied for bank loan. 86% of the total loan issued were good loans while 13% were bad loans. The bank to fund the loan with £436M and in return, received £473M as loan repayment.